Rename the class into acceleration operations
