+++ public/components/datasources/components/manage/accelerations/acceleration_operation.tsx
